Update ui5-support-lib.js
Adds the possibility to add the boolean data.preventBrowserHistory option to the manifest
If set to true, navigating inside the same route will prevent the browser history

We use routes that build up on selecting different options. If we click back on the browser arrow, we need to click multiple times until we reach the inital start point e.g. Dashboard again

if the flag is not set, there are no side effects to current behaviours